4-fasta-embedder/requirements.txt
pandas~=2.2.2
transformers~=4.43.3
python-dotenv~=1.0.1
tqdm~=4.66.5
numpy~=1.26
biopython~=1.84